+++ generators/app/templates/api/controllers/AuthController.js
@@ -81,6 +81,16 @@ module.exports = {
     },
 
     /**
+     * Google Plus authorization\linking
+     */
+    google: function (req, res) {
+        passport.authenticate('jwt', function (error, user) {
+            req.user = user;
+            passport.authenticate('google-plus-token', _onPassportAuth.bind(this, req, res))(req, res);
+        })(req, res);
+    },
+
+    /**
      * Server ping
      * Useful when need to check if it's server is down or some logic is break
      */